add constraint
