+++ pyro/distributions/transforms/unit_cholesky.py
@@ -6,20 +6,6 @@ from torch.distributions import constraints
 from torch.distributions.transforms import Transform
 
 
-class _UnitLowerCholesky(torch.distributions.Constraint):
-    """
-    Constrain to lower-triangular square matrices with all ones diagonals.
-    """
-    event_dim = 2
-
-    def check(self, value):
-        value_tril = value.tril()
-        lower_triangular = (value_tril == value).view(value.shape[:-2] + (-1,)).min(-1)[0]
-
-        ones_diagonal = (value.diagonal(dim1=-2, dim2=-1) == 1).min(-1)[0]
-        return lower_triangular & ones_diagonal
-
-
 class UnitLowerCholeskyTransform(Transform):
     """
     Transform from unconstrained matrices to lower-triangular matrices with
@@ -39,6 +25,5 @@ class UnitLowerCholeskyTransform(Transform):
 
 
 __all__ = [
-    "_UnitLowerCholesky",
     "UnitLowerCholeskyTransform",
 ]
